refactor: 优化动画与伤害数值渲染逻辑，提升性能与可维护性
1.  重构动画完成检测逻辑，新增totalDuration预计算并简化AnimationEngine代码
2.  为SelectorConfig添加预解析间隔缓存，移除运行时正则匹配
3.  统一DamageAccumulator使用double类型存储总伤害，修复类型不一致问题
4.  优化HUD渲染流程，减少重复配置获取和同步开销
5.  简化ActiveDamageNumber的isComplete实现，使用轻量完成检查
6.  修复轨迹列表截断逻辑，优化内存占用

+++ forge-1.20.1/src/main/java/com/stylizeddamage/forge/client/TotalDamageHudRenderer.java
@@ -83,7 +83,7 @@ public final class TotalDamageHudRenderer {
     // ── Animation state ─────────────────────────────────────────────
     private TotalPhase totalPhase = TotalPhase.INACTIVE;
     private int totalAnimTick;
-    private float prevTotalDamage;
+    private double prevTotalDamage;
     private int bounceTick;
     private final List<TrailAnimState> trailAnimStates = new ArrayList<>();
 
@@ -623,7 +623,7 @@ public final class TotalDamageHudRenderer {
         }
     }
 
-    private static double calculateFontSize(TotalDamageConfig config, float totalDamage) {
+    private static double calculateFontSize(TotalDamageConfig config, double totalDamage) {
         double steps = totalDamage / 100.0;
         double raw = config.baseFontSize() + steps * config.sizeOffsetPerThousand();
         return Math.max(config.baseFontSize(), Math.min(raw, config.sizeOffsetMax()));
